requirements.txt
pydglab-ws
qrcode
python-osc
colorlog
pillow
pyyaml
psutil
PySide6
qasync
websockets
pyee>=8.2.2